insurance_scanner/requirements.txt
flask==3.0.3
flask-cors==4.0.1
easyocr==1.7.1
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
scikit-learn==1.4.2
watchdog==4.0.1
joblib==1.4.0
scipy==1.13.0
